Clean up support bots endpoint
