+++ wp_api/src/wp_com/client.rs
@@ -3,9 +3,10 @@ use super::endpoint::jetpack_connection_endpoint::{
 };
 use super::endpoint::oauth2::{Oauth2RequestBuilder, Oauth2RequestExecutor};
 use super::endpoint::subscribers::{SubscribersRequestBuilder, SubscribersRequestExecutor};
-use super::endpoint::support_bots::SupportBotsRequestExecutor;
+use super::endpoint::support_bots_endpoint::{
+    SupportBotsRequestBuilder, SupportBotsRequestExecutor,
+};
 use crate::api_client_generate_request_builder;
-use crate::wp_com::endpoint::support_bots::SupportBotsRequestBuilder;
 use crate::{
     ParsedUrl, WpApiClientDelegate, api_client_generate_api_client,
     api_client_generate_endpoint_impl, auth::WpAuthenticationProvider,
