+++ wp_api/src/wp_com/endpoint.rs
@@ -1,4 +1,4 @@
 pub mod jetpack_connection_endpoint;
 pub mod oauth2;
 pub mod subscribers;
-pub mod support_bots;
+pub mod support_bots_endpoint;
